requirements.txt
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
greenlet==3.0.1
idna==3.4
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.2
pandas==2.1.2
pathspec==0.11.2
platformdirs==3.11.0
protobuf==4.21.12
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
TA-Lib==0.4.25
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
